extract data processing
